installed cookie-parser, fixed routes/signup.js
